requirements.txt
aiogram==2.9
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv~=0.14.0
qiwipy
SQLAlchemy~=1.3.18
gino~=1.0.1
asyncpg~=0.21.0